+++ index.html
@@ -1,15 +1,15 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <meta name="theme-color" content="#08090A" />
-        <title>oluwasefunmi</title>
-        <link rel="stylesheet" href="./source/styles/tailwind.css" />
-        <link rel="stylesheet" href="./source/styles/base.css" />
-    </head>
-    <body>
-        <div id="app"></div>
-        <script type="module" src="./source/main.ts"></script>
-    </body>
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta name="theme-color" content="#08090A" />
+    <title>oluwasefunmi</title>
+    <link rel="stylesheet" href="./source/styles/tailwind.css" />
+    <link rel="stylesheet" href="./source/styles/base.css" />
+  </head>
+  <body>
+    <div id="app"></div>
+    <script type="module" src="./source/app/main.ts"></script>
+  </body>
 </html>
