+++ source/features/playground/components/ProductDetails.tsx
@@ -83,12 +83,12 @@ export const ProductDetails = (props: ProductDetailsProps) => {
             <div class="flex flex-wrap gap-2 md:gap-3">
               {For(product.colors, (color) => {
                 const isSelected = Cell.derived(
-                  () => selectedColor.get() === color,
+                  () => selectedColor.get() === color
                 );
                 const optionClasses = Cell.derived(() =>
                   isSelected.get()
                     ? "py-2 px-4 rounded-full text-sm text-white/90 bg-white/15 border border-white/40 transition-all duration-200 ease-out cursor-pointer flex items-center gap-2"
-                    : "py-2 px-4 rounded-full text-sm text-white/70 bg-white/5 border border-transparent transition-all duration-200 ease-out cursor-pointer hover:bg-white/10 hover:border-white/10 flex items-center gap-2",
+                    : "py-2 px-4 rounded-full text-sm text-white/70 bg-white/5 border border-transparent transition-all duration-200 ease-out cursor-pointer hover:bg-white/10 hover:border-white/10 flex items-center gap-2"
                 );
 
                 return (
