+++ source/features/playground/experiments/BlurCarousel.module.css
