+++ source/pages/PortfolioHome.tsx
@@ -1,9 +1,9 @@
 import { Link } from "retend/router";
 import type { RouteComponent } from "retend/router";
-import { LayeredCard } from "../components/LayeredCard";
+import { LayeredCard } from "@/components/ui/LayeredCard";
 import classes from "./PortfolioHome.module.css";
-import { CurrentPageTitle } from "@/components/CurrentPageTitle";
-import { StarShower } from "@/components/StarShower";
+import { CurrentPageTitle } from "@/components/layout/CurrentPageTitle";
+import { StarShower } from "@/components/ui/StarShower";
 
 const PortfolioHome: RouteComponent = () => {
   return (
